// frontend/src/app/types.ts
export type ManagedCompany = {
  id: string;
  name: string;
  segment: string;
  createdAt: string;
};